src directory reorganization
