--- my-app/src/scenarioObservation.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-import AthleteMap from './athletemap';
-
-class NewObservation extends Component {
-    state = {  }
-    render() { 
-        return ( <AthleteMap/> );
-    }
-}
- 
-export default NewObservation;
