--- my-app/src/scenarioROM.jsx
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-class NewROM extends Component {
-    state = {  }
-    render() { 
-        return ( <h4>Range of Motion</h4> );
-    }
-}
- 
-export default NewROM;
